refactor(dev): use entry point to bootstrap the app
